pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easygame"
version = "0.1.0"
description = "A lightweight 2D game framework"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Ilya Kamenshchikov", email = "ilya.kamenshchikov@gmail.com" },
]
keywords = ["game", "framework", "2d", "sprite", "pygame", "pyglet"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]
dependencies = [
    "Pillow>=12.1.1",
    "pyglet>=2.1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.15",
    "mypy>=1.19",
]

[project.urls]
Repository = "https://github.com/easygame/easygame"

# ---------------------------------------------------------------------------
# Ruff — linter & formatter
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
# Default rule set (E, F, W) matches the existing clean baseline.
# UP031 is explicitly selected to catch %-format stragglers.
select = ["E", "F", "W", "UP031"]
ignore = []

[tool.ruff.lint.per-file-ignores]
# Tests may use assertions, private access, and broader imports.
"tests/**" = ["S101"]

# ---------------------------------------------------------------------------
# mypy — static type checking
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
